+++ bevy_ios_safearea/src/plugin.rs
@@ -1,11 +1,24 @@
 use bevy::prelude::*;
-#[cfg(target_os = "ios")]
-use bevy::window::PrimaryWindow;
 
+/// Struct providing iOS device safe area insets.
 #[derive(Copy, Clone, Debug, Default)]
 pub struct IosSafeArea {
+    /// The inset from the top of the screen.
+    ///
+    /// This value accounts for elements like the notch or status bar.
     pub top: f32,
+    /// The inset from the bottom of the screen.
+    ///
+    /// This value accounts for elements like the home indicator.
     pub bottom: f32,
+    /// The inset from the left side of the screen.
+    ///
+    /// This value is non-zero for devices with rounded corners or unique screen shapes.
+    pub left: f32,
+    /// The inset from the right side of the screen.
+    ///
+    /// This value is non-zero for devices with rounded corners or unique screen shapes.
+    pub right: f32,
 }
 
 #[derive(Resource, Clone, Debug, Default)]
@@ -20,6 +33,14 @@ impl IosSafeAreaResource {
 
     pub fn bottom(&self) -> f32 {
         self.safe_area.map_or(0., |a| a.bottom)
+    }
+
+    pub fn left(&self) -> f32 {
+        self.safe_area.map_or(0., |a| a.left)
+    }
+
+    pub fn right(&self) -> f32 {
+        self.safe_area.map_or(0., |a| a.right)
     }
 }
 
@@ -38,7 +59,7 @@ impl Plugin for IosSafeAreaPlugin {
 #[cfg(target_os = "ios")]
 fn init(
     windows: NonSend<bevy::winit::WinitWindows>,
-    window_query: Query<Entity, With<PrimaryWindow>>,
+    window_query: Query<Entity, With<bevy::window::PrimaryWindow>>,
     mut res: ResMut<IosSafeAreaResource>,
 ) {
     use bevy::utils::tracing;
@@ -54,8 +75,15 @@ fn init(
 
             let top = unsafe { crate::native::swift_safearea_top(ui_view) };
             let bottom = unsafe { crate::native::swift_safearea_bottom(ui_view) };
+            let left = unsafe { crate::native::swift_safearea_left(ui_view) };
+            let right = unsafe { crate::native::swift_safearea_right(ui_view) };
 
-            let safe_area = IosSafeArea { top, bottom };
+            let safe_area = IosSafeArea {
+                top,
+                bottom,
+                left,
+                right,
+            };
 
             tracing::debug!("safe area updated: {:?}", safe_area);