+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* support `left`&`right` safearea [#9](https://github.com/rustunit/bevy_ios_safearea/pull/9)
+ 
 ## [0.1.6] - 2025-01-15
 
 ### Fixed